Fix UserModel/User creation controller

+++ server/controller/userController.js
@@ -1,36 +1,46 @@
-import UserModel from '../models/user.js';
 import mongoose from 'mongoose';
 
+import UserModel from '../models/user.js';
+
+const ObjectId = mongoose.Types.ObjectId;
+
 const UserController = {
-  getUserById: async (req, res) => {
+  getUserById: async (req) => {
     try {
-      const id = req.params === undefined ? req.id : req.params.id;
-      const result = UserModel.findById(id);
-      return result ? result : {};
+      var id = req.params === undefined ? req.id : req.params.id;
+      if (mongoose.isValidObjectId(id)) {
+        const result = await UserModel.findById(id);
+        return result;
+      } else {
+        return {statusCode: 400, error: 'Bad Request', message: 'Invalid object ID'};
+      }
     } catch (error) {
-      throw new Error(error);
+      return new Error(error);
     }
   },
-  getUsers: async (req, res) => {
+  getUsers: async () => {
     try {
-      return await UserModel.find();
+      const result = await UserModel.find();
+      return result;
     } catch (error) {
-      throw new Error(error);
+      return new Error(error);
     }
   },
   addUser: async (req, res) => {
     try {
-      const body = req.body;
+      const body = {... req.body, _id: new ObjectId()};
+      console.log("Body: " , body)
       const newUser = new UserModel(body);
       const existing = await UserModel.findOne({ username: body.username });
 
       if (existing) {
-        return new Error('User already exists');
+        return {statusCode: 400, error: 'Bad Request', message: 'User already exists'};
       } else {
-        return await newUser.save();
+        const result = await newUser.save();
+        return result;
       }
     } catch (error) {
-      throw new Error(error);
+      return new Error(error);
     }
   },
   editUser: async (req, res) => {
@@ -39,14 +49,15 @@ const UserController = {
       const body = req.body;
       const found = UserModel.findOne({ id: id });
 
-      if (found) {
-        return await UserModel.findByIdAndUpdate(id, body, { new: true });
+      if (mongoose.isValidObjectId(id) && found) {
+        const result = await UserModel.findByIdAndUpdate(id, body, { new: true });
+        return result;
       } else {
-        return new Error('User does not exist');
+        return {statusCode: 400, error: 'Bad Request', message: 'User does not exist or ObjectId is invalid'};
       }
 
     } catch (error) {
-      throw new Error(error);
+      return new Error(error);
     }
   },
   deleteUser: async (req, res) => {
@@ -54,14 +65,15 @@ const UserController = {
       const id = req.params === undefined ? req.id : req.params.id;
       const found = UserModel.findOne({ id: id });
 
-      if (found) {
-        return await UserModel.findByIdAndRemove(id);
+      if (mongoose.isValidObjectId(id) && found) {
+        const result = await UserModel.findByIdAndRemove(id);
+        return result;
       } else {
-        return new Error('User does not exist');
+        return {statusCode: 400, error: 'Bad Request', message: 'User does not exist or ObjectId is invalid'};
       }
 
     } catch (error) {
-      throw new Error(error);
+      return new Error(error);
     }
   },
 };
